DOe maar niet

+++ TheHunt/TheHunt/Form1.cs
@@ -21,8 +21,6 @@ namespace TheHunt
         public static Size startRes = new Size();
         MediaPlayer bgm = new MediaPlayer();
         MediaPlayer klikMP = new MediaPlayer();
-
-
         public Form1()
         {
             InitializeComponent();
@@ -66,7 +64,7 @@ namespace TheHunt
             }
         }
 
-        //speel geluid af als het niet gemute is
+
         public void speelKlikGeluid(object sender, EventArgs e)
         {
             if (!isMuted)
@@ -91,28 +89,20 @@ namespace TheHunt
             bgm.Play();
             }
 
-        private void GoFullscreen(bool gofullscreen)
+        private void GoFullscreen(bool fullscreen)
         {
-
-            if (gofullscreen)
+            if (fullscreen)
             {
-                //Fullscreen activeren
                 this.WindowState = FormWindowState.Normal;
                 this.Bounds = Screen.PrimaryScreen.Bounds;
                 initOptionsPanel();
-                buttonFuSc.Text = "Full Screen: On";
-                goFullScreen = false;
             }
             else
             {
-                //Fullscreen deactiveren
                 this.ClientSize = new Size(startRes.Width, startRes.Height);
                 this.Location = new Point((Screen.PrimaryScreen.WorkingArea.Width - startRes.Width) / 2, (Screen.PrimaryScreen.WorkingArea.Height - startRes.Height) / 2);
                 initOptionsPanel();
-                buttonFuSc.Text = "Full Screen: Off";
-                goFullScreen = true;
             }
-            initOptionsPanel();
         }
 
         private void Form1_Load(object sender, EventArgs e)
@@ -122,24 +112,42 @@ namespace TheHunt
                 Directory.CreateDirectory(Directory.GetCurrentDirectory() + "/SFX");
             }
 
-            if (!File.Exists(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav") || !File.Exists(Directory.GetCurrentDirectory() + "/SFX/bgm.wav"))
+            if (!File.Exists(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav") || !File.Exists(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav"))
             {
                 MemoryStream memStream = new MemoryStream();
 
                 if (!File.Exists(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav"))
                 {
-                    Properties.Resources.klikgeluid.CopyTo(memStream);
-                    byte[] byteArray = memStream.ToArray();
-                    File.WriteAllBytes(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav", byteArray);
-                    memStream.SetLength(0);
+
+                Properties.Resources.klikgeluid.CopyTo(memStream);
+                byte[] byteArray = memStream.ToArray();
+                File.WriteAllBytes(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav", byteArray);
+                memStream.SetLength(0);
                 }
-                if (!File.Exists(Directory.GetCurrentDirectory() + "/SFX/bgm.wav"))
+                else if (!File.Exists(Directory.GetCurrentDirectory() + "/SFX/bgm.wav"))
+                {
+                Properties.Resources.bgm.CopyTo(memStream);
+                    byte[] byteArray = memStream.ToArray();
+                File.WriteAllBytes(Directory.GetCurrentDirectory() + "/SFX/bgm.wav", byteArray);
+                memStream.SetLength(0);
+                }
+                else if (!File.Exists(Directory.GetCurrentDirectory() + "/SFX/bgm.wav") && !File.Exists(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav"))
                 {
                     Properties.Resources.bgm.CopyTo(memStream);
                     byte[] byteArray = memStream.ToArray();
                     File.WriteAllBytes(Directory.GetCurrentDirectory() + "/SFX/bgm.wav", byteArray);
                     memStream.SetLength(0);
-                }
+
+                    Properties.Resources.klikgeluid.CopyTo(memStream);
+                    byteArray = memStream.ToArray();
+                    File.WriteAllBytes(Directory.GetCurrentDirectory() + "/SFX/klikgeluid.wav", byteArray);
+                    memStream.SetLength(0);
+
+            }
+                else
+            {
+                    //doe niks
+            }   
             }
 
 
@@ -161,7 +169,7 @@ namespace TheHunt
 
         public void initOptionsPanel()
         {
-            pictureBox1.Visible = false;
+            pictureBox4.Visible = false;
             pictureBox2.Visible = false;
             pictureBox3.Visible = false;
             pictureBox4.Visible = false;
@@ -172,8 +180,8 @@ namespace TheHunt
             pictureBox6.Size = Properties.Resources.uitleg.Size;
             pictureBox6.Location = new Point(panel1.Location.X + panel1.Width / 2, panel1.Location.Y);
 
-            panel1.Width = (int)(this.Width * 0.85);
-            panel1.Height = (int)(this.Height * 0.75);
+            panel1.Width = (int)(this.Width * 0.8);
+            panel1.Height = (int)(this.Height * 0.7);
             panel1.Location = new Point((int)(this.Width * 0.1), (int)(this.Height * 0.1));
